Add files via upload
If you open my project , you will see the product image on the home page.
Clicking on the product Name will take  you to the product list(i'm using routing method).
The product list will have a button on the ViewDetails.
If you click on the View Details button the detail of the product will appear.
i'm craeting json,productss.ts,services.I'm using internal Api.
Open Api not used because open Api does not work on my laptop 
i tried to more time because of that i did internal Api. 
Add on the json file are automatically on the home page(i'm using service concept)
add on the the home page are automatically on the product list(using @Input) i did not finish this completely.